Added MovingPlatform Script
Place Moving platform script on the game object you would like to have
move in a loop. Use the boolean switches "vertical" and "horizontal" to
set the direction of the movement. Set the Amplitude to control the
force and set speed between 0 - 1 to get the best effect. Though we
should keep with the default settings.