ParentToPlatform Script Added
After attaching the "MovingPlatform" script to the platform you would
like to move, create an empty game object as a child to the platform.
Place a box collider component on the child and set it to isTrigger.
Raise the box collider slightly above the platforms collider so the
collider on the player will enter it. The script will make the Player a
child of the platform while the Player is within the trigger collider.