Added Respawn and Checkpoint functionality
Place the Respawn.js on any gameObject where the collider kills the
player.
Create an empty gameObject called SpawnPoint, drag this onto the
appropriate script.
Add the Checkpoint.js to gameObject with a isTrigger boxCollider. When
the player collides the spawnPoint will move to this position. The
Player will now respawn here.