Pause Screen, Death Screen, Victory Screen, also added audio to scene.
Look to the changes I've made and read the comments in the scripts to
understand how things work with the screens, all GUI screens must be
turned on in the scene before building the executable. This is because
the GameObject needs to be active in order to be found by the Start
functions. I will supply you all with prefabs for each screen instance.